자료구조 HW1

+++ 데이터구조/과제1/1-1.소수판별.c
@@ -1,13 +1,14 @@
 #include <stdio.h>
+
 void is_prime(int n) {
 	int i = 2;
-	for (i = 2; i <= (int)(n / 2); i++) {
+	for (i = 2; i <= (n / 2); i++) {
 		if (n % i == 0) {
 			printf("composite");
             break;
 		}
 
-        if (i == (int)(n/2)){
+        if (i == (n/2)){
             printf("prime");
         }   
 	}
